+++ esp-hal/CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `AnyPin` now implements `From<GpioPin<N>>`. (#2326)
+- Added `degrade` function for SPI instances (SPI2, SPI3) to obtain `AnySpi`. (#2334)
+
 ### Changed
+
+- Peripheral type erasure for SPI (#2334)
 
 ### Fixed